requirements.txt
asgiref==3.8.1
channels==4.2.0
channels-redis==4.2.0
Django==5.1.4
djangorestframework==3.15.2
gunicorn==21.2.0
pillow==11.0.0
psycopg2-binary==2.9.10
redis==5.0.1
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.1
django-storages==1.14.2
boto3==1.34.7
uvicorn==0.25.0
